+++ packages/terra-application-navigation/src/common/_PopupMenuListItem.jsx
@@ -95,7 +95,6 @@ const PopupMenuListItem = ({
 
   /* eslint-disable jsx-a11y/no-noninteractive-tabindex */
   /* eslint-disable jsx-a11y/no-noninteractive-element-interactions */
-  /* eslint-disable jsx-a11y/no-noninteractive-element-to-interactive-role */
   return (
     <li
       {...ariaSpread}