+++ packages/terra-slide-panel-manager/src/terra-dev-site/doc/example/SlidePanelManagerAccessoryExample.jsx
@@ -289,4 +289,3 @@ const SlidePanelManagerExample = () => (
 );
 
 export default SlidePanelManagerExample;
-/* eslint-disable max-classes-per-file */