+++ packages/terra-slide-panel-manager/src/terra-dev-site/doc/example/SlidePanelManagerExample.jsx
@@ -284,4 +284,3 @@ const SlidePanelManagerExample = () => (
 );
 
 export default SlidePanelManagerExample;
-/* eslint-disable max-classes-per-file */