+++ packages/terra-application-layout/CHANGELOG.md
@@ -5,6 +5,7 @@ Unreleased
 ----------
 ### Changed
 * Minor file changes to comply with eslint rules
+* Removed unused disable eslint directives
 
 5.38.0 - (June 16, 2020)
 ------------------