+++ packages/terra-application-utility/CHANGELOG.md
@@ -3,6 +3,8 @@ Changelog
 
 Unreleased
 ----------
+### Changed
+* Remove unused disable eslint directives
 
 2.35.0 - (June 9, 2020)
 ------------------