+++ packages/terra-date-time-picker/CHANGELOG.md
@@ -5,6 +5,7 @@ Unreleased
 ----------
 ### Changed
 * Minor file changes to comply with eslint rules
+* Remove unused disable eslint directives
 
 4.39.0 - (June 16, 2020)
 ------------------