+++ packages/terra-time-input/CHANGELOG.md
@@ -5,6 +5,7 @@ Unreleased
 ----------
 ### Changed
 * Minor file changes to comply with eslint rules
+* Remove unused disable eslint directives
 
 4.29.0 - (June 9, 2020)
 ------------------